project/services/web/requirements.txt
flask==1.1.1
psycopg2-binary==2.8.3